refactor(GCFLogger-test-helpers): move module from test/fixtures to test/
