+++ packages/gcf-utils/test/GCFLogger-test-helpers.ts
